CommentForm component is in its own file
